chore: Move all go code to the test
To make explicit that this it not a go module anymore and the users are
expected to use the bash scripts directly.

The go code is only used for testing of this repository.

Ticket: QA-524
