+++ license_test.go
@@ -23,9 +23,115 @@ import (
 	"path/filepath"
 	"testing"
 
+	"errors"
+	"strings"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+const packageLocation string = "github.com/mendersoftware/mendertesting"
+
+var known_license_files []string = []string{}
+
+// Specify a license file for a dependency explicitly, avoiding the check for
+// common license file names.
+func SetLicenseFileForDependency(license_file string) {
+	known_license_files = append(known_license_files, "--add-license="+license_file)
+}
+
+var firstEnterpriseCommit = ""
+
+// This should be set to the oldest commit that is not part of Open Source, only
+// part of Enterprise, if any. IOW it should be the very first commit after the
+// fork point, on the Enterprise branch.
+func SetFirstEnterpriseCommit(sha string) {
+	firstEnterpriseCommit = sha
+}
+
+func CheckMenderCompliance(t *testing.T) {
+	t.Run("Checking Mender compliance", func(t *testing.T) {
+		err := checkMenderCompliance()
+		assert.NoError(t, err)
+	})
+}
+
+type MenderComplianceError struct {
+	Output string
+	Err    error
+}
+
+func (m *MenderComplianceError) Error() string {
+	return fmt.Sprintf("MenderCompliance failed with error: %s\nOutput: %s\n", m.Err, m.Output)
+}
+
+func checkMenderCompliance() error {
+	pathToTool, err := locatePackage()
+	if err != nil {
+		return err
+	}
+
+	args := []string{path.Join(pathToTool, "check_license_source_files.sh")}
+	if firstEnterpriseCommit != "" {
+		args = append(args, "--ent-start-commit", firstEnterpriseCommit)
+	}
+	cmd := exec.Command("bash", args...)
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		return &MenderComplianceError{
+			Err:    err,
+			Output: string(output),
+		}
+	}
+
+	args = []string{path.Join(pathToTool, "check_commits.sh")}
+	cmd = exec.Command("bash", args...)
+	output, err = cmd.CombinedOutput()
+	if err != nil {
+		return &MenderComplianceError{
+			Err:    err,
+			Output: string(output),
+		}
+	}
+
+	args = []string{path.Join(pathToTool, "check_license.sh")}
+	args = append(args, known_license_files...)
+	cmd = exec.Command("bash", args...)
+	output, err = cmd.CombinedOutput()
+	if err != nil {
+		return &MenderComplianceError{
+			Err:    err,
+			Output: string(output),
+		}
+	}
+	return nil
+}
+
+func locatePackage() (string, error) {
+	finalpath := path.Join("vendor", packageLocation)
+	_, err := os.Stat(finalpath)
+	if err == nil {
+		return finalpath, nil
+	}
+
+	gopath := os.Getenv("GOPATH")
+	if gopath == "" {
+		return "", errors.New("Cannot check for licenses if " +
+			"mendertesting is not vendored and GOPATH is unset.")
+	}
+
+	paths := strings.Split(gopath, ":")
+	for i := 0; i < len(paths); i++ {
+		finalpath = path.Join(paths[i], "src", packageLocation)
+		_, err := os.Stat(finalpath)
+		if err == nil {
+			return finalpath, nil
+		}
+	}
+
+	return "", fmt.Errorf("Package '%s' could not be located anywhere in GOPATH (%s)",
+		packageLocation, gopath)
+}
 
 func resetKnownLicenses() {
 	known_license_files = []string{}
